+++ ui/components/combobox/RELEASENOTES.md
@@ -9,7 +9,9 @@
 ### Changed
 
 - Increase font-size and height of combobox height on a touch device
--
+- Increase height of a listbox option on a touch device
+- Changed the size of icon in a plain listbox to inherit the size of the `body`
+
 ### Fixed
 
 - Resolve issue with deprecated inline listbox not displaying correctly when multiple pills forces the container to grow