+++ ui/components/menus/RELEASENOTES.md
@@ -10,6 +10,7 @@
 
 - On touch devices, set tap target size of menu row item to `$height-tappable`
 - On touch devices, set `font-size` of menu to `$font-size-5`
+- Remove `font-size` declaration from `slds-dropdown__header` to inherit `font-size` from `body` element
 
 ## 2.8.0